move class to utils
